requirements.txt
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-extensions==4.1
django-sslserver==0.22
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
MarkupSafe==3.0.2
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
sqlparse==0.5.3
typing_extensions==4.14.1
Werkzeug==3.1.3
